Fix ContourConnection to SaddleConnection conversion

+++ libflatsurf/benchmark/flat_triangulation_combinatorial.benchmark.cc
@@ -34,7 +34,7 @@ using namespace flatsurf::test;
 void FlatTriangulationCombinatorialFlip(State& state) {
   auto L = makeLCombinatorial();
 
-  const int N = L->size();
+  const int N = static_cast<int>(L->size());
 
   int idx = 0;
